+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project (more or less) adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.3.0] - 2026-01-14
+
+### Added:
+- ScriptFolderManager Screen added
+- Adding and removing Scriptfolders within the TUI
+
 ## [0.2.0] - 2026-01-12
 
 ### Added: